+++ common/src/main/java/net/mehvahdjukaar/supplementaries/common/entities/IFluteParrot.java
@@ -1,0 +1,8 @@
+package net.mehvahdjukaar.supplementaries.common.entities;
+
+import net.minecraft.world.entity.player.Player;
+
+public interface IFluteParrot {
+
+    void supplementaries$setPartyByFlute(Player player);
+}