+++ common/src/main/java/net/mehvahdjukaar/supplementaries/mixins/LevelRendererMixin.java
@@ -1,12 +1,18 @@
 package net.mehvahdjukaar.supplementaries.mixins;
 
 import com.mojang.blaze3d.vertex.PoseStack;
+import com.mojang.datafixers.util.Either;
 import net.mehvahdjukaar.supplementaries.client.renderers.items.SlingshotRendererHelper;
+import net.mehvahdjukaar.supplementaries.common.network.ClientReceivers;
 import net.minecraft.client.Camera;
 import net.minecraft.client.Minecraft;
+import net.minecraft.client.multiplayer.ClientLevel;
 import net.minecraft.client.renderer.GameRenderer;
 import net.minecraft.client.renderer.LevelRenderer;
 import net.minecraft.client.renderer.LightTexture;
+import net.minecraft.core.BlockPos;
+import net.minecraft.world.level.Level;
+import org.jetbrains.annotations.Nullable;
 import org.joml.Matrix4f;
 import org.spongepowered.asm.mixin.Final;
 import org.spongepowered.asm.mixin.Mixin;
@@ -32,5 +38,17 @@ public abstract class LevelRendererMixin {
     private void supplementaries$renderSlingshotOutline(PoseStack matrixStack, float partialTicks, long finishNanoTime, boolean blockOutlines, Camera camera, GameRenderer renderer, LightTexture lightTexture, Matrix4f matrix4f, CallbackInfo ci) {
         if (blockOutlines) SlingshotRendererHelper.renderBlockOutline(matrixStack, camera, this.minecraft);
     }
+
+    @Shadow
+    @Nullable
+    private ClientLevel level;
+
+
+    @Inject(method = "notifyNearbyEntities", at = @At("HEAD"))
+    private void setPartying(Level worldIn, BlockPos pos, boolean isPartying, CallbackInfo info) {
+      ClientReceivers.setDisplayParrotsPartying(worldIn, Either.right(pos), isPartying);
+    }
+
+
 
 }